tools/m8t/requirements.txt
chess==1.10.0
python-chess==1.999
PyYAML==6.0.1
